requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
architect==0.6.0
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncstdlib==3.10.9
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl-cffi==0.5.10
Django==4.1.13
django-postgres-extra==2.0.8
django-redis==5.4.0
djongo==1.3.6
dnspython==2.4.2
duckduckgo-search==3.9.9
elastic-transport==8.10.0
elasticsearch==7.17.0
exceptiongroup==1.2.0
Faker==20.1.0
fastapi==0.104.1
Flask==3.0.0
frozenlist==1.4.0
g4f==0.1.9.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
mmh3==4.0.1
multidict==6.0.4
neo4j==5.12.0
neobolt==1.7.17
neomodel==5.2.0
nest-asyncio==1.5.8
numpy==1.26.2
outcome==1.3.0.post0
Pillow==10.1.0
platformdirs==4.0.0
psycopg2==2.9.9
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyExecJS==1.5.1
pyjsparser==2.7.1
pymongo==4.6.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
redis==5.0.1
requests==2.31.0
selenium==4.15.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
sqlparse==0.2.4
starlette==0.27.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==1.26.18
uvicorn==0.24.0.post1
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.3
